fix: Change the data type of `validatorIndex` from `uint256` to `uint64` in the `ValidatorData` struct

+++ l1-contracts/src/structs/ValidatorData.sol
@@ -3,6 +3,6 @@ pragma solidity >=0.8.0 <0.9.0;
 
 struct ValidatorData {
     address eigenPod;
-    uint256 validatorIndex;
+    uint64 validatorIndex;
     bytes delegatePubKey;
 }
